[test] move catch main into catch2 dir and introduce test main target
