--- test/test_main.cpp
@@ -1,30 +1,0 @@
-// SPDX-License-Identifier: MIT
-// SPDX-FileCopyrightText: 2023 NeoFOAM authors
-
-#include <catch2/catch_session.hpp>
-#include <catch2/catch_test_macros.hpp>
-#include <catch2/generators/catch_generators_adapters.hpp>
-#include <Kokkos_Core.hpp>
-
-#include "NeoFOAM/core/mpi/environment.hpp"
-
-int main(int argc, char* argv[])
-{
-    NeoFOAM::mpi::MPIInit mpi(argc, argv);
-
-    // Initialize Catch2
-    Kokkos::initialize(argc, argv);
-    Catch::Session session;
-
-    // Specify command line options
-    int returnCode = session.applyCommandLine(argc, argv);
-    if (returnCode != 0) // Indicates a command line error
-        return returnCode;
-
-    int result = session.run();
-
-    // Run benchmarks if there are any
-    Kokkos::finalize();
-
-    return result;
-}
